+++ client/containers/DevTools/DevTools.js
@@ -1,7 +1,8 @@
 import React from 'react';
-import { createDevTools } from 'redux-devtools';
+import {createDevTools} from 'redux-devtools';
 import LogMonitor from 'redux-devtools-log-monitor';
 import DockMonitor from 'redux-devtools-dock-monitor';
+import {persistState} from 'redux-devtools'
 
 const DevTools = createDevTools(
   <DockMonitor toggleVisibilityKey='ctrl-h'
@@ -12,5 +13,13 @@ const DevTools = createDevTools(
   </DockMonitor>
 )
 
-export default DevTools
-module.exports = DevTools
+function getDebugSessionKey() {
+  // You can write custom logic here!
+  // By default we try to read the key from ?debug_session=<key> in the address bar
+  const matches = window.location.href.match(/[?&]debug_session=([^&]+)\b/);
+  return (matches && matches.length > 0)? matches[1] : null;
+}
+
+exports.DevTools = DevTools  // DevTools ui component
+exports.persistState = persistState
+exports.getDebugSessionKey = getDebugSessionKey